// types.ts
export interface ServiceItem {
  id: string;
  title: string;
  description: string;
  icon: string;
  image: string;
}

export interface BenefitItem {
  id: string;
  title: string;
  description: string;
  icon: string;
}

export interface PartnerItem {
  id: string;
  name: string;
  logo: string;
  description: string;
}

export interface FAQItem {
  question: string;
  answer: string;
}